requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.0.3
joblib==1.3.2
scikit-learn==1.2.2
tensorflow==2.12.0
mysql-connector-python==8.0.33
Pillow==8.2.0
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1+cpu
numpy==1.23.5
werkzeug==2.3.7
pandas==1.2.4